chore: restore size_t and add debug function

+++ libft/ft_index_of.c
@@ -6,15 +6,15 @@
 /*   By: sblanco- <sblanco-@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/09 20:32:28 by sblanco-          #+#    #+#             */
-/*   Updated: 2024/11/10 09:29:48 by sblanco-         ###   ########.fr       */
+/*   Updated: 2024/11/10 14:21:01 by sblanco-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_index_of(char *str, char c)
+size_t	ft_index_of(char *str, char c)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str && str[i])
@@ -23,5 +23,5 @@ int	ft_index_of(char *str, char c)
 			return (i);
 		i++;
 	}
-	return (-1);
+	return ((size_t)-1);
 }
